merge master into presto-timestamp

+++ tests/integration/src/test/java/org/apache/pulsar/tests/integration/docker/ContainerExecResultBytes.java
@@ -26,7 +26,7 @@ import lombok.Data;
 @Data(staticConstructor = "of")
 public class ContainerExecResultBytes {
 
-    private final int exitCode;
+    private final long exitCode;
     private final byte[] stdout;
     private final byte[] stderr;
